update to new version and add new packages

+++ deb_repo/thirdreality-hacore-config/build.sh
@@ -30,7 +30,11 @@ version=$(grep '^Version:' ${current_dir}/DEBIAN/control | awk '{print $2}')
 print_info "Version: $version"
 
 if [[ "$CLEAN" == true ]]; then
-    
+
+    rm -rf ${output_dir}/var > /dev/null 2>&1
+    rm -rf ${output_dir}/etc > /dev/null 2>&1
+    rm -rf ${output_dir}/usr > /dev/null 2>&1
+
     rm -rf "${output_dir}" > /dev/null 2>&1
 
     print_info "Removing ${config_dir}"
@@ -44,6 +48,11 @@ fi
 
 if [[ "$REBUILD" == true ]]; then
     print_info "hacore-config_${version}.deb rebuilding ..."
+
+    rm -rf ${output_dir}/var > /dev/null 2>&1
+    rm -rf ${output_dir}/etc > /dev/null 2>&1
+    rm -rf ${output_dir}/usr > /dev/null 2>&1
+
     rm -rf "${output_dir}" > /dev/null 2>&1
     rm -rf "${config_dir}" > /dev/null 2>&1
 
@@ -61,34 +70,37 @@ mkdir -p "${config_dir}"
 print_info "syncing DEBIAN ..."
 cp ${current_dir}/DEBIAN ${output_dir}/ -R
 
-
 if [ ! -d "${config_dir}/homeassistant" ]; then
+    print_info "Copying homeassistant files to ${config_dir} ..."
     cp ${current_dir}/homeassistant/*  "${config_dir}/" -R
 fi
 
-rm -rf ${output_dir}/var > /dev/null 2>&1
-rm -rf ${output_dir}/etc > /dev/null 2>&1 
-rm -rf ${output_dir}/usr > /dev/null 2>&1
+if [ ! -d "${output_dir}/var/lib/homeassistant" ]; then
+    rm -rf ${output_dir}/var > /dev/null 2>&1
+    rm -rf ${output_dir}/etc > /dev/null 2>&1 
+    rm -rf ${output_dir}/usr > /dev/null 2>&1
 
-mkdir -p ${output_dir}/var/lib/homeassistant
-chmod 755 ${output_dir}/var/lib/homeassistant
+    mkdir -p ${output_dir}/var/lib/homeassistant
+    chmod 755 ${output_dir}/var/lib/homeassistant
 
-mkdir -p ${output_dir}/etc
-chmod 755 ${output_dir}/etc
+    mkdir -p ${output_dir}/etc
+    chmod 755 ${output_dir}/etc
 
-ls -l ${config_dir}
+    ls -l ${config_dir}
 
-cp ${config_dir}/* ${output_dir}/var/lib/homeassistant/ -R
-rm -rf ${output_dir}/var/lib/homeassistant/homeassistant/.HA_VERSION
+    print_info "Copying homeassistant files from ${config_dir} to ${output_dir} ..."
 
-cp /etc/hassio.json ${output_dir}/etc/ 
+    cp ${config_dir}/* ${output_dir}/var/lib/homeassistant/ -R
+    rm -rf ${output_dir}/var/lib/homeassistant/homeassistant/.HA_VERSION
+
+    if [ -f "/etc/hassio.json" ]; then
+        cp /etc/hassio.json ${output_dir}/etc/ 
+    else
+        print_info "No /etc/hassio.json found, skipping copy."
+    fi
+fi
 
 print_info "Start to build hacore-config_${version}.deb ..."
 dpkg-deb --build ${output_dir} ${current_dir}/hacore-config_${version}.deb
-
-
-rm -rf ${output_dir}/var > /dev/null 2>&1
-rm -rf ${output_dir}/etc > /dev/null 2>&1
-rm -rf ${output_dir}/usr > /dev/null 2>&1
 
 print_info "Build hacore-config_${version}.deb finished ..."
